packages/db/pyproject.toml
[project]
name = "hephae-db"
version = "0.1.0"
description = "All Firestore, BigQuery, and GCS data access for hephae-forge"
requires-python = ">=3.12"
dependencies = [
    "hephae-common",
    "firebase-admin>=6.6.0",
    "google-cloud-bigquery>=3.27.0",
    "google-cloud-storage>=2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["hephae_db"]
